requirements.txt
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
pytest==8.1.2
tomli==2.0.1
